requirements.txt
pandas==1.5.2
requests==2.28.2
google-cloud-bigquery[pandas]==3.2.0